+++ src/mongo/db/exec/geo_near.cpp
@@ -447,7 +447,7 @@ GeoNear2DStage::GeoNear2DStage(const GeoNearParams& nearParams,
       _boundsIncrement(0.0) {
     _specificStats.keyPattern = twoDIndex->keyPattern();
     _specificStats.indexName = twoDIndex->indexName();
-    _specificStats.indexVersion = twoDIndex->version();
+    _specificStats.indexVersion = static_cast<int>(twoDIndex->version());
 }
 
 
@@ -745,7 +745,7 @@ GeoNear2DSphereStage::GeoNear2DSphereStage(const GeoNearParams& nearParams,
       _boundsIncrement(0.0) {
     _specificStats.keyPattern = s2Index->keyPattern();
     _specificStats.indexName = s2Index->indexName();
-    _specificStats.indexVersion = s2Index->version();
+    _specificStats.indexVersion = static_cast<int>(s2Index->version());
 
     // initialize2dsphereParams() does not require the collator during the GEO_NEAR_2DSPHERE stage.
     // It only requires the collator for index key generation. For query execution,
